Make filter construction work for sizes other than 1
The program claimed to support arbitrary filter sizes, but in reality the
matrix of coefficients was hard-coded for a radius of 1. Specifying
larger sizes lead to GPU aborts due to the kernel running off the end
of the array when trying to fetch coefficients.

Replace the hardcoded matrix with one created on the fly according to
https://stackoverflow.com/questions/8204645/implementing-gaussian-blur-how-to-calculate-convolution-matrix-kernel.